Moved pure functions to presenter file.
